Build optimization 10.06.2017 (#2037)
* Refactor

* Refactor
