+++ core-java/src/main/java/com/baeldung/serialization/Address.java
@@ -1,4 +1,4 @@
-package com.baeuldung.serialization;
+package com.baeldung.serialization;
 
 public class Address {
