+++ core-java/src/main/java/com/baeldung/serialization/Person.java
@@ -1,4 +1,4 @@
-package com.baeuldung.serialization;
+package com.baeldung.serialization;
 import java.io.Serializable;
 
 public class Person implements Serializable {
