+++ core-java/src/test/java/com/baeldung/serialization/PersonUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeuldung.serialization;
+package com.baeldung.serialization;
 
 import static org.junit.Assert.assertTrue;
