+++ spring-5-mvc/src/test/java/com/baeldung/Spring5ApplicationIntegrationTest.java
@@ -1,11 +1,9 @@
-package com.baeldung.test;
+package com.baeldung;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import com.baeldung.Spring5Application;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = Spring5Application.class)
